+++ gateways/MeprCreditCardGateway.php
@@ -3,9 +3,9 @@
  * Credit card gateway alias
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2018 Pronamic
+ * @copyright 2005-2019 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\MemberPress
  */
 
-class_alias( 'Pronamic\WordPress\Pay\Extensions\MemberPress\CreditCardGateway', 'MeprCreditCardGateway' );
+class_alias( 'Pronamic\WordPress\Pay\Extensions\MemberPress\Gateways\CreditCardGateway', 'MeprCreditCardGateway' );