+++ gateways/MeprDirectDebitIDealGateway.php
@@ -3,9 +3,9 @@
  * Direct Debit mandate via iDEAL gateway alias
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2018 Pronamic
+ * @copyright 2005-2019 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\MemberPress
  */
 
-class_alias( 'Pronamic\WordPress\Pay\Extensions\MemberPress\DirectDebitIDealGateway', 'MeprDirectDebitIDealGateway' );
+class_alias( 'Pronamic\WordPress\Pay\Extensions\MemberPress\Gateways\DirectDebitIDealGateway', 'MeprDirectDebitIDealGateway' );