+++ src/MemberPress.php
@@ -3,7 +3,7 @@
  * MemberPress
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2018 Pronamic
+ * @copyright 2005-2019 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\MemberPress
  */
@@ -11,6 +11,7 @@
 namespace Pronamic\WordPress\Pay\Extensions\MemberPress;
 
 use MeprTransaction;
+use MeprOptions;
 
 /**
  * WordPress pay MemberPress
@@ -34,5 +35,17 @@ class MemberPress {
 		}
 
 		return ( $transaction->status === $status );
+	}
+
+	/**
+	 * Get currency.
+	 *
+	 * @return string
+	 */
+	public static function get_currency() {
+		$mepr_options = MeprOptions::fetch();
+
+		// @link https://gitlab.com/pronamic/memberpress/blob/1.2.4/app/models/MeprOptions.php#L136-137
+		return $mepr_options->currency_code;
 	}
 }